fix(canvas): ctx.font 归一化——em 相对字号/数值 weight/small-caps/族名校验 (R34xx)
WPT 2d.text.font.* 全过（element + offscreen worker 双面）：
- parse_css_with_current：em 相对当前字号（'1em' 默认 10px → 10px）；rem 恒 16px
- 数值 weight（100-900）保留（'italic 300 12px serif' round-trip）
- small-caps variant 建模（getter 重建 'small-caps'）
- getFont 归一化：CSV 感知拆分（引号内逗号/转义）+ 通用族关键字小写 + 自定义名保留
- family 标识符校验（{} / ; 等拒绝 + CSS-wide 关键字拒绝——'10px {bogus}'/'initial' 无效）
- canvas 761 / engine 2109 / wpt-runner 171 全绿；clippy 零警告

+++ crates/canvas/src/path_tests.rs
@@ -224,7 +224,7 @@ fn shadow_blur_softens_edge_r3240() {
         } else {
             ctx.fill_rect(5.0, 5.0, 10.0, 10.0);
         }
-        ctx.get_image_data(px, py, 1, 1).data[3]
+        ctx.get_image_data(px as i32, py as i32, 1, 1).data[3]
     };
 
     // (21,15)：silhouette 右边沿（20）外 1px，不被红矩形覆盖（到 15）。
@@ -297,7 +297,7 @@ fn shadow_blur_multipass_convex_falloff_r3242() {
     ctx.fill_rect(10.0, 10.0, 10.0, 10.0);
 
     // 读右边沿外 d=1,2,3,4 的阴影 alpha（y=15 矩形内部，x=21..24）。
-    let alpha_at = |x: u32| ctx.get_image_data(x, 15, 1, 1).data[3] as i32;
+    let alpha_at = |x: u32| ctx.get_image_data(x as i32, 15, 1, 1).data[3] as i32;
     let a1 = alpha_at(21);
     let a2 = alpha_at(22);
     let a3 = alpha_at(23);
